Update paloalto-software-upgrade.sh
unset all variables so that the loop would work across multiple firewalls and updated an if statement that broke during the loop.  Can confirm that it will loop across firewalls and restart the questions again successfully.  Next version will need to simply apply certain variables globally (version to download and to install / download only) so that the script just runs without interaction.