Create paloalto-dynupdates.sh
Used for auditing (only) of dynamic updates across firewalls.  Dependency on using paloalto-configs.sh to download the device state, extract the running-config.xml and then grab all the dynamic update configs we are interested in.  This is an old script and does not have the mail function in it, but it's easy to add if you want.  Ye old copy and paste will do most of it if you need it.