Update paloalto-software-upgrade.sh
I found that I was doing the rssc AFTER checking if it's the latest version, which clearly doesn't work.  Simply moved rssc up above the if 'force' loop. Also updated the instructions for the force, because I was looking at $1 instead of $2.  If I left it at $1 it wouldn't find 'force' in inventory and exit.